requirements.txt
absl-py==0.9.0
appdirs==1.4.4
astor==0.8.1
attrs==22.2.0
audioread==2.1.9
Automat==22.10.0
brotlipy==0.7.0
cachetools==4.0.0
certifi==2021.10.8
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
chardet==3.0.4
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.0.3
cloudpickle==2.1.0
colorama==0.4.5
Comet==3.1.0
comet-ml==3.31.21
configobj==5.0.6
configparser==5.2.0
constantly==15.1.0
contextlib2==21.6.0
cryptography @ file:///tmp/build/80754af9/cryptography_1652101588893/work
cycler==0.11.0
decorator==4.4.2
dina==0.1
docker-pycreds==0.4.0
dulwich==0.20.50
et-xmlfile==1.1.0
everett==3.1.0
facenet-pytorch==2.5.2
ffmpeg==1.4
filelock==3.7.1
fonttools==4.25.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.26
google-auth==1.11.2
google-auth-oauthlib==0.4.1
grpcio==1.27.2
h5py==3.6.0
hyperlink==21.0.0
hyperopt==0.1.2
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.19.2
imageio-ffmpeg==0.4.8
importlib-metadata==4.10.1
incremental==22.10.0
joblib==1.1.0
json-tricks==3.15.5
jsonpatch==1.25
jsonpointer==2.0
jsonschema==4.17.3
kiwisolver==1.4.2
librosa==0.9.1
llvmlite==0.36.0
lxml==4.9.2
Markdown==3.2.1
matplotlib==3.5.1
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
moviepy==1.0.3
munkres==1.1.4
networkx==2.8.2
nni==2.7
numba==0.53.0
numpy==1.22.3
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.5.64
openpyxl==3.0.10
packaging==21.3
pandas==1.3.5
pathtools==0.1.2
Pillow==8.4.0
pooch==1.6.0
prettytable==3.3.0
proglog==0.1.10
promise==2.3
protobuf==3.19.4
psutil==5.9.0
pthflops==0.4.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydub==0.25.1
pymongo==4.1.1
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
PythonWebHDFS==0.2.3
pytz==2021.3
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==18.1.1
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.0
requests-toolbelt==0.10.1
resampy==0.2.2
responses==0.21.0
rsa==4.0
schema==0.7.5
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
semantic-version==2.10.0
sentry-sdk==1.5.4
shortuuid==1.0.8
simplejson==3.17.6
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
smmap==5.0.0
SoundFile==0.10.3.post1
sox==1.4.1
subprocess32==3.5.4
tensorboard==2.1.0
termcolor==1.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.5.4
torch==1.10.2
torchaudio==0.10.2
torchfile==0.1.0
torchsummary==1.5.1
torchvision==0.11.2
tornado==6.1
tqdm==4.62.3
Twisted==22.10.0
typeguard==2.13.3
typing_extensions==4.0.1
urllib3 @ file:///opt/conda/conda-bld/urllib3_1650639997961/work
visdom==0.1.8.9
wandb==0.12.9
wcwidth==0.2.5
websocket-client==0.57.0
websockets==10.3
Werkzeug==1.0.0
wrapt==1.14.1
wurlitzer==3.0.3
yaspin==2.1.0
zipp==3.7.0
zope.interface==5.5.2
